Change uint to uint256 for variable types

+++ WEEK1/FINALTASK~SimpleStorage.sol
@@ -3,23 +3,23 @@
  pragma solidity ^0.8.18;
 
 contract SimpleStorage{
-    uint myNumber;
+    uint256 myNumber;
 
    struct Person{   
        string name;
-       uint Number;
+       uint256 Number;
     }
-   function set(uint _number) public virtual{
+   function set(uint256 _number) public virtual{
        myNumber=_number;
     
     }
      
-      function get() public view returns(uint){
+      function get() public view returns(uint256){
      return myNumber;
      }
     
 }
 // DEPLOYED LOCALLY IN ANVIL BLOCKCHAIN...
 //Deployer: 0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266
-//Deployed to: 0x610178dA211FEF7D417bC0e6FeD39F05609AD788
-//Transaction hash: 0x29783b9cb2cfa5cb19904a31fa4895c7e3f3ee91ef23ea95320a70b58062d0be
+//Deployed to: 0x5FbDB2315678afecb367f032d93F642f64180aa3
+//Transaction hash: 0xff75ddda06d1e1e586d1be14a995d361ff03e9cf2ead40b5ff2c945e213a3f7f
